🎨 move model and services from main to core folder
